# springboot-swagger-elastic/src/main/webapp/swagger2.json
{
	"swagger": "2.0",
	"info": {
		"description": "This is a my swagger.",
		"version": "1.0.0",
		"title": "Swagger Collection",
		"termsOfService": "http://swagger.io/terms/",
		"contact": {
			"email": "michael.wang@mw.io"
		},
		"license": {
			"name": "Apache 2.0",
			"url": "http://www.apache.org/licenses/LICENSE-2.0.html"
		}
	},
	"host": "localhost:8080",
	"basePath": "/",
...
	"schemes": [
		"http"
	],
	"paths": {
...
		"/greeting": {
			"get": {
				"tags": [
					"greeting"
				],
				"summary": "Greeting",
				"description": "Returns a single greeting",
				"operationId": "greeting",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"name": "name",
						"in": "query",
						"description": "name",
						"required": false,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "successful operation",
						"schema": {
							"$ref": "#/definitions/Greeting"
						}
					},
					"400": {
						"description": "Invalid Greeting supplied"
					},
					"404": {
						"description": "Greeting not found"
					}
				}
			}
		},
...
		"/user": {
			"post": {
				"tags": [
					"user"
				],
				"summary": "Create user",
				"description": "This can only be done by the logged in user.",
				"operationId": "createUser",
				"produces": [
					"application/xml",
					"application/json"
				],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"description": "Created user object",
						"required": true,
						"schema": {
							"$ref": "#/definitions/User"
						}
					}
				],
				"responses": {
					"default": {
						"description": "successful operation"
					}
				}
			}
		},
...
	},
	"securityDefinitions": {
		"petstore_auth": {
			"type": "oauth2",
			"authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
			"flow": "implicit",
			"scopes": {
				"write:pets": "modify pets in your account",
				"read:pets": "read your pets"
			}
		},
		"api_key": {
			"type": "apiKey",
			"name": "api_key",
			"in": "header"
		}
	},
	"definitions": {
...
		"User": {
			"type": "object",
			"properties": {
				"id": {
					"type": "integer",
					"format": "int64"
				},
				"username": {
					"type": "string"
				},
				"firstName": {
					"type": "string"
				},
				"lastName": {
					"type": "string"
				},
				"email": {
					"type": "string"
				},
				"password": {
					"type": "string"
				},
				"phone": {
					"type": "string"
				},
				"userStatus": {
					"type": "integer",
					"format": "int32",
					"description": "User Status"
				}
			},
			"xml": {
				"name": "User"
			}
		},
...
		"Greeting": {
			"properties": {
				"id": {
					"description": "Greeting Id",
					"type": "integer"
				},
				"content": {
					"description": "content",
					"type": "string"
				}
			}
		},
...
	},
...
}